Add files via upload
MAX6675 supports an SCK up to 4.3 MHz. This means that the minimum duration of an SCK clock can be 200ns. 
I integrated the following methods into the library to reduce reading time from the MAX6675:
- FastReadCelcius ()
- FastSpiRead ()
With the use of this routine the reading times are reduced from about 33 ms. to about 250 us, 
therefore reducing this phase of about 130 times.